tokens moved to separate auth store; other stores moved to appropriate files
