+++ front/src/stores/NavigationStore.js
@@ -1,0 +1,15 @@
+import {action, observable} from "mobx";
+
+export default class NavigationStore {
+  @observable rootStore;
+  @observable drawerStatus = false;
+
+  constructor(rootStore) {
+    this.rootStore = rootStore;
+  }
+
+  @action.bound
+  toggleDrawer(){
+    this.drawerStatus = !this.drawerStatus;
+  }
+}
